02-openapi-agent/requirements.txt
azure-ai-projects[agents]
azure-identity
python-dotenv
jsonref
